requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
asn1crypto==0.24.0
beautifulsoup4==4.7.1
chardet==3.0.4
cryptography==2.6.1
distro-info==0.21
Django==3.0.8
django-cors-headers==3.4.0
djangorestframework==3.11.0
entrypoints==0.3
graphviz==0.14
html5lib==1.0.1
keyring==17.1.1
keyrings.alt==3.1.1
lxml==4.3.2
numpy==1.19.1
pandas==1.1.0
psycopg2==2.8.5
pycrypto==2.6.1
pycurl==7.43.0.2
pydot==1.4.1
pydot3==1.0.9
PyGObject==3.30.4
pyparsing==2.4.7
python-apt==1.8.4.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
pyxattr==0.6.1
pyxdg==0.25
SecretStorage==2.3.1
six==1.15.0
soupsieve==1.8
sqlparse==0.3.1
unattended-upgrades==0.1
virtualenv==15.1.0
webencodings==0.5.1
youtube-dl==2019.1.17
